+++ src/lib/components/action-msg/SingleActionMsg.tsx
@@ -1,3 +1,5 @@
+import { observer } from "mobx-react-lite";
+
 import { useSingleActionMsgProps } from "lib/hooks/useSingleMessageProps";
 import type { Message } from "lib/types";
 
@@ -9,17 +11,14 @@ interface SingleActionMsgProps {
   success: boolean;
   singleMsg?: boolean;
 }
-export const SingleActionMsg = ({
-  messages,
-  type,
-  success,
-  singleMsg,
-}: SingleActionMsgProps) => {
-  const singleMsgProps = useSingleActionMsgProps(
-    type,
-    success,
-    messages,
-    singleMsg
-  );
-  return <SingleMsg {...singleMsgProps} />;
-};
+export const SingleActionMsg = observer(
+  ({ messages, type, success, singleMsg }: SingleActionMsgProps) => {
+    const singleMsgProps = useSingleActionMsgProps(
+      type,
+      success,
+      messages,
+      singleMsg
+    );
+    return <SingleMsg {...singleMsgProps} />;
+  }
+);
