--- src/lib/pages/contract-details/components/tables/index.ts
@@ -1,3 +1,0 @@
-export * from "./migration";
-export * from "./RelatedProposalsTable";
-export * from "./TxsTable";
